# openapi.yaml
openapi: 3.0.3
info:
  title: Helix Health Portal API
  description: |
    RESTful API for the Helix Health Portal outpatient patient portal.
    Used as the sample application for the QA automation course.
  version: 1.0.0
  contact:
    name: Helix Health Portal
servers:
  - url: http://localhost:4000/api
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiSuccess:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
    ApiError:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Invalid credentials"
...
    Appointment:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: integer
        providerId:
          type: integer
        appointmentTypeId:
          type: integer
        appointmentTypeName:
          type: string
          example: Follow-up
        scheduledAt:
          type: string
          format: date-time
        durationMinutes:
          type: integer
          example: 30
        status:
          type: string
          enum: [scheduled, confirmed, completed, cancelled, no_show]
        location:
          type: string
          example: Room 101
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
...
paths:
...
  /appointments:
    get:
      tags: [Appointments]
      summary: List appointments (filtered by role — provider sees all, patient sees own)
      operationId: listAppointments
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [scheduled, confirmed, completed, cancelled, no_show]
        - name: providerId
          in: query
          schema:
            type: integer
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Appointment list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Appointment'
  /patients/{patientId}/appointments:
    get:
      tags: [Appointments]
      summary: List appointments for a specific patient
      operationId: listPatientAppointments
      security:
        - bearerAuth: []
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Patient appointment list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Appointment'
        '404':
          description: Patient not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

    post:
      tags: [Appointments]
      summary: Book an appointment for a patient
      operationId: createAppointment
      security:
        - bearerAuth: []
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [appointmentTypeId, providerId, scheduledAt]
              properties:
                appointmentTypeId:
                  type: integer
                providerId:
                  type: integer
                scheduledAt:
                  type: string
                  format: date-time
                location:
                  type: string
                notes:
                  type: string
      responses:
        '201':
          description: Appointment created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Appointment'
        '409':
          description: Scheduling conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /appointments/{appointmentId}:
    get:
      tags: [Appointments]
      summary: Get a single appointment
      operationId: getAppointment
      security:
        - bearerAuth: []
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Appointment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccess'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

    patch:
      tags: [Appointments]
      summary: Update appointment status or details
      operationId: updateAppointment
      security:
        - bearerAuth: []
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [confirmed, cancelled, no_show, completed]
                notes:
                  type: string
                location:
                  type: string
      responses:
        '200':
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccess'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /appointments/availability:
    get:
      tags: [Appointments]
      summary: Get available time slots for a provider
      operationId: getAvailability
      security:
        - bearerAuth: []
      parameters:
        - name: providerId
          in: query
          required: true
          schema:
            type: integer
        - name: appointmentTypeId
          in: query
          required: true
          schema:
            type: integer
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Available slots
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccess'

  # ── Medical Records ───────────────────────────────────────────────────────────
...